moved functions for reading/writing models to separate module
