+++ src/lib.rs
@@ -1,5 +1,6 @@
 //! Command line tool to show clear git graphs arranged for your branching model.
 
+pub mod config;
 pub mod graph;
 pub mod print;
 pub mod settings;
